遠回りな実装を回避して不要なContextを削除
